+++ app_run.go
@@ -2,6 +2,7 @@ package newrelic
 
 import (
 	"encoding/json"
+	"strings"
 	"time"
 
 	"github.com/newrelic/go-agent/internal"
@@ -16,6 +17,9 @@ type appRun struct {
 	// the security policies.
 	AttributeConfig *internal.AttributeConfig
 	Config          Config
+
+	// firstAppName is the value of Config.AppName up to the first semicolon.
+	firstAppName string
 }
 
 func newAppRun(config Config, reply *internal.ConnectReply) *appRun {
@@ -91,6 +95,9 @@ func newAppRun(config Config, reply *internal.ConnectReply) *appRun {
 	if run.Config.DistributedTracer.Enabled {
 		run.Config.CrossApplicationTracer.Enabled = false
 	}
+
+	// Cache the first application name set on the config
+	run.firstAppName = strings.SplitN(config.AppName, ";", 2)[0]
 
 	if "" != run.Reply.RunID {
 		js, _ := json.Marshal(settings(run.Config))